Move around MerkleTree* structs
